Converted Lablib to work with OS 10.12 release
All projects coverted.
More stringent type casting for long to int in many places.
A few NS constant names have been deprecated (but replaced)
enums must now be “typedef enum” with a name declared
Definitions of variables in headers must now be explicitly “extern”

+++ Plugins/Fixate/FTXTController.m
@@ -34,7 +34,7 @@ NSString *FTTrialWindowZoomKey = @"FTTrialWindowZoom";
     
     zoomValue = [[sender selectedCell] tag];
     [self setScaleFactor:zoomValue / 100.0];
-    [[NSUserDefaults standardUserDefaults] setObject:[NSNumber numberWithInt:zoomValue] 
+    [[NSUserDefaults standardUserDefaults] setObject:[NSNumber numberWithInt:(int)zoomValue]
                 forKey:FTTrialWindowZoomKey];
 }
